Odds and Ends
Ignore-this: 3288940a047840a290a97f019dc5155b

darcs-hash:20120515140132-33929-1aac8057d2def4f6b4b44836da1f28bb3dc9a86f

+++ src/main/scala/de/dimond/tippspiel/model/Game.scala
@@ -10,10 +10,10 @@ import net.liftweb.http.S
 
 object Game {
   def all = games.values.toList.sortWith((e1, e2) => (e1.date compareTo e2.date) < 0)
-  private var games: Map[Int, Game] = Map()
+  private var games: Map[Long, Game] = Map()
 }
 
-case class Game(id: Int, teamHome: TeamReference, teamAway: TeamReference, date: DateTime, location: Location) {
+case class Game(id: Long, teamHome: TeamReference, teamAway: TeamReference, date: DateTime, location: Location) {
   Game.games = Game.games + ((id, this))
 }
